requirements.txt
agno>=2.4.8
chromadb>=1.4.1
google-genai>=1.62.0
ollama>=0.6.1
openai>=2.17.0
python-dotenv>=1.2.1
sqlalchemy>=2.0.46